+++ NEWS.md
@@ -3,6 +3,9 @@
 * Add `split()` methods to divide into groups.
 * Add `hist()` methods to plot ILR histograms.
 
+## Enhancements
+* `transform_ilr()` gained a new argument allowing the specify the base of the transformation.
+
 # nexus 0.1.0
 
 * First release.