+++ NEWS.md
@@ -1,4 +1,4 @@
-# nexus 0.1.0.9000
+# nexus 0.2.0
 ## New classes and methods
 * Add `hist()` methods to plot ILR histograms.
 * Add `quantile()` methods for compositional data matrix.